move amount to separate module to make future refactoring simpler
lint fixes
