+++ rust/src/command/mod.rs
@@ -5,10 +5,7 @@ pub mod store;
 use crate::{
     block::{precomputed::PrecomputedBlock, BlockHash},
     command::signed::{SignedCommand, SignedCommandWithKind},
-    ledger::{
-        account::{Amount, Nonce},
-        public_key::PublicKey,
-    },
+    ledger::{account::Nonce, amount::Amount, public_key::PublicKey},
     protocol::serialization_types::staged_ledger_diff as mina_rs,
 };
 use log::trace;
